+++ plock/config.py
@@ -21,6 +21,8 @@ BUILDOUT_OPT = (
     'buildout:eggs-directory=eggs-directory',
     '-U',)
 
+EGGS_TOTAL = 235  # Number of eggs in working set
+
 SEARCH_OPER = 'AND'
 SEARCH_SPEC = {'description': 'plone', 'keyword': 'plone', 'summary': 'plone'}